votingapp/voteApp/requirements.txt
Django==1.8.6
Flask==0.10.1
Flask-Bootstrap==3.3.5.3
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==3.0.0
Py4J==0.6
PyJWT==1.4.0
PyYAML==3.11
SQLAlchemy==1.0.11
Werkzeug==0.10.4
argparse==1.2.1
django-filter==0.13.0
djangorestframework==3.3.3
djoser==0.4.3
dominate==2.1.12
fuzzywuzzy==0.5.0
itsdangerous==0.24
mysqlclient==1.3.7
nltk==3.1
numpy==1.10.2
oauthlib==1.0.3
pandas==0.17.1
pbr==1.8.1
progressbar==2.3
pylint-django==0.7.1
pylint-plugin-utils==0.2.3
pyparsing==2.0.6
python-Levenshtein==0.12.0
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
rest-social-auth==0.4.2
six==1.10.0
stevedore==1.10.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wsgiref==0.1.2
